producer/requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
deprecation==2.1.0
idna==3.7
jwcrypto==1.5.6
kafka-python==2.0.2
packaging==24.0
psycopg2==2.9.9
pycparser==2.22
python-keycloak==4.0.0
requests==2.32.2
requests-toolbelt==1.0.0
typing_extensions==4.12.0
urllib3==2.2.1
